requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
async-timeout==5.0.1
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
colorama==0.4.6
coverage==7.9.1
cryptography==45.0.4
Deprecated==1.2.18
dnspython==2.7.0
exceptiongroup==1.3.0
Faker==37.4.0
fastapi==0.115.14
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
limits==5.4.0
motor==3.7.1
multidict==6.6.2
packaging==25.0
pamqp==3.3.0
pluggy==1.6.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.7.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.13.2
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
bcrypt==4.1.3
passlib==1.7.4
PyJWT==2.10.1